feat: backend dos requisitos 20 ao 32

+++ back-end/src/database/migrations/20211110001208-create-sales.js
@@ -25,7 +25,7 @@ module.exports = {
         foreignKey: true,
       },
       totalPrice: {
-        type: Sequelize.INTEGER,  /* DECIMAL(10,2), */
+        type: Sequelize.DECIMAL(9,2),  /* DECIMAL(10,2), */
         field: 'total_price'
       },
       deliveryAddress: {
